+++ src/schema_voyager/html/components/value_type.cljs
@@ -1,23 +1,17 @@
 (ns schema-voyager.html.components.value-type
-  (:require [schema-voyager.html.util :as util]))
+  (:require
+   [schema-voyager.html.util :as util]))
 
-(defn tuple-composition [{:keys [db/tupleAttrs db/tupleType db/tupleTypes]}]
-  (cond
-    tupleAttrs :db.type.tuple/composite
-    tupleType  :db.type.tuple/homogeneous
-    tupleTypes :db.type.tuple/heterogeneous))
-
-(defn angle-span [child]
+(defn- angle-span [child]
   [:span "< " child " >"])
 
-(defn angle-list [children]
+(defn- angle-list [children]
   [angle-span [util/comma-list children]])
 
 (def ^:private one-abbr [:abbr {:title ":db.cardinality/one"} "one"])
 (def ^:private many-abbr [:abbr {:title ":db.cardinality/many"} "many"])
 (def ^:private tuple-abbr [:abbr {:title ":db.type/tuple"} "tuple"])
 (def ^:private tuples-abbr [:abbr {:title ":db.type/tuple"} "tuples"])
-(def ^:private reference-abbr [:abbr {:title ":db.type/ref"} "reference"])
 (def ^:private references-abbr [:abbr {:title ":db.type/ref"} "references"])
 (def ^:private composed-abbr [:abbr {:title ":db/tupleAttrs"} "composed"])
 (def ^:private homogeneous-abbr [:abbr {:title ":db/tupleType"} "homogeneous"])
@@ -25,132 +19,96 @@
 (defn- keyword-abbr [kw]
   [:abbr {:title (pr-str kw)} (name kw)])
 
-(defn- tuple-attrs-list [{:keys [db/tupleAttrs]}]
+(defn- links-or-abbr [value-type]
+  (if (keyword? value-type)
+    [keyword-abbr value-type]
+    [util/coll-links value-type]))
+
+(defn- pipe-links-or-abbr [value-type]
+  (if (keyword? value-type)
+    [keyword-abbr value-type]
+    [util/coll-pipe-links value-type]))
+
+(defn- tuple-attrs-list [attrs]
   [angle-list (map (fn [attr]
                      ^{:key (:db/ident attr)}
                      [util/attr-link attr])
-                   tupleAttrs)])
+                   attrs)])
 
-(defn- heterogeneous-types-list [{:keys [db/tupleTypes db.schema/tuple-references]}]
-  (let [refs-by-position (zipmap (map :db.schema.tuple/position tuple-references)
-                                 (map :db.schema/references tuple-references))]
-    [angle-list
-     (map-indexed (fn [position value-type]
-                    ^{:key position}
-                    (let [refs (get refs-by-position position)]
-                      (if (and (= :db.type/ref value-type)
-                               refs)
-                        [util/coll-pipe-links refs]
-                        [keyword-abbr value-type])))
-                  tupleTypes)]))
+(defn- heterogeneous-types-list [value-types]
+  [angle-list
+   (map-indexed (fn [position value-type]
+                  ^{:key position}
+                  [pipe-links-or-abbr value-type])
+                value-types)])
 
-(defn- dispatch [{:keys [db/valueType db.schema/references db/tupleType] :as attribute}]
-  (let [tuple?     (= :db.type/tuple valueType)
-        ref?       (= :db.type/ref valueType)
-        tuple-ref? (= :db.type/ref tupleType)]
-    [(cond tuple? (tuple-composition attribute)
-           ref?   valueType
-           :else  ::default)
-     (if (and (seq references)
-              (or ref? tuple-ref?))
-       :referred
-       ::default)]))
+(defn- is-or-references [value-type]
+  (if (keyword? value-type) "is" references-abbr))
 
-(defmulti p
-  (fn [{:keys [db/cardinality] :as attribute}]
-    (into [cardinality] (dispatch attribute))))
+(defn- has-or-references [value-type]
+  (if (keyword? value-type) "has" references-abbr))
 
-;; composite tuples
-(defmethod p [:db.cardinality/one :db.type.tuple/composite ::default] [attribute]
-  [:p
-   "This attribute is " one-abbr " " tuple-abbr ", " composed-abbr " of other attributes "
-   [tuple-attrs-list attribute]". "
-   "It's managed by Datomic, and should not be set manually."])
-(defmethod p [:db.cardinality/many :db.type.tuple/composite ::default] [attribute]
+(defmulti p :db/valueType)
+(defmulti p-composite :db/cardinality)
+(defmulti p-heterogeneous :db/cardinality)
+(defmulti p-homogeneous :db/cardinality)
+(defmulti p-basic :db/cardinality)
+
+(defmethod p :db.type/tuple.composite [attribute] [p-composite attribute])
+(defmethod p :db.type/tuple.homogeneous [attribute] [p-homogeneous attribute])
+(defmethod p :db.type/tuple.heterogeneous [attribute] [p-heterogeneous attribute])
+(defmethod p :default [attribute] [p-basic attribute])
+
+(defmethod p-composite :db.cardinality/many [{:keys [db/tupleAttrs]}]
   ;; NOTE: In practice a composite tuple is probably never card many.
   [:p
    "This attribute is " many-abbr " " tuples-abbr ", each " composed-abbr " of other attributes "
-   [tuple-attrs-list attribute]". "
+   [tuple-attrs-list tupleAttrs]". "
+   "It's managed by Datomic, and should not be set manually."])
+(defmethod p-composite :default [{:keys [db/tupleAttrs]}]
+  [:p
+   "This attribute is " one-abbr " " tuple-abbr ", " composed-abbr " of other attributes "
+   [tuple-attrs-list tupleAttrs]". "
    "It's managed by Datomic, and should not be set manually."])
 
-;; homogeneous tuples
-(def ^:private homogeneous-one-span [:span
-                                     "This attribute has " one-abbr " value, a " homogeneous-abbr " " tuple-abbr " of variable length. "
-                                     "Each member of the tuple "])
-(def ^:private homogeneous-many-span [:span
-                                      "This attribute has " many-abbr " values, " homogeneous-abbr " " tuples-abbr " of variable length. "
-                                      "Each member of each tuple "])
-(defn- homogeneous-referred-span [{:keys [db.schema/references]}]
-  [:span references-abbr " one " [util/coll-links references] "."])
-(defn- homogeneous-unreferred-span [{:keys [db/tupleType]}]
-  [:span "is one " [keyword-abbr tupleType] "."])
-
-(defmethod p [:db.cardinality/one :db.type.tuple/homogeneous :referred] [attribute]
-  [:p homogeneous-one-span [homogeneous-referred-span attribute]])
-(defmethod p [:db.cardinality/one :db.type.tuple/homogeneous ::default] [attribute]
-  [:p homogeneous-one-span [homogeneous-unreferred-span attribute]])
-(defmethod p [:db.cardinality/many :db.type.tuple/homogeneous :referred] [attribute]
-  [:p homogeneous-many-span [homogeneous-referred-span attribute]])
-(defmethod p [:db.cardinality/many :db.type.tuple/homogeneous ::default] [attribute]
-  [:p homogeneous-many-span [homogeneous-unreferred-span attribute]])
-
-;; heterogeneous tuples
-(defn- heterogeneous-one-span [{:keys [db/tupleTypes]}]
-  [:span "This attribute has " one-abbr " value, a " heterogeneous-abbr " " tuple-abbr " of length "
-   (count tupleTypes) ". The tuple is of type "])
-(defn- heterogeneous-many-span [{:keys [db/tupleTypes]}]
-  [:span "This attribute has " many-abbr " values, " heterogeneous-abbr " " tuples-abbr " of length "
-   (count tupleTypes) ". Each tuple is of type "])
-(defmethod p [:db.cardinality/one :db.type.tuple/heterogeneous ::default] [attribute]
+(defmethod p-homogeneous :db.cardinality/many [{:keys [db/tupleType]}]
   [:p
-   [heterogeneous-one-span attribute]
-   [heterogeneous-types-list attribute]])
-(defmethod p [:db.cardinality/many :db.type.tuple/heterogeneous ::default] [attribute]
+   "This attribute has " many-abbr " values, " homogeneous-abbr " " tuples-abbr " of variable length. "
+   "Each member of each tuple " (is-or-references tupleType) " one "
+   [links-or-abbr tupleType]])
+(defmethod p-homogeneous :default [{:keys [db/tupleType]}]
   [:p
-   [heterogeneous-many-span attribute]
-   [heterogeneous-types-list attribute]])
+   "This attribute has " one-abbr " value, a " homogeneous-abbr " " tuple-abbr " of variable length. "
+   "Each member of the tuple " (is-or-references tupleType) " one "
+   [links-or-abbr tupleType]])
 
-;; references
-(defmethod p [:db.cardinality/one :db.type/ref :referred] [{:keys [db.schema/references]}]
-  [:p "This attribute has " one-abbr " value, a " reference-abbr " to one " [util/coll-links references] "."])
-(defmethod p [:db.cardinality/one :db.type/ref ::default] [_attribute]
-  [:p "This attribute " references-abbr " " one-abbr " value."])
-(defmethod p [:db.cardinality/many :db.type/ref :referred] [{:keys [db.schema/references]}]
-  [:p "This attribute has " many-abbr " values, each a " reference-abbr " to one " [util/coll-links references] "."])
-(defmethod p [:db.cardinality/many :db.type/ref ::default] [_attribute]
-  [:p "This attribute " references-abbr " " many-abbr " values."])
+(defmethod p-heterogeneous :db.cardinality/many [{:keys [db/tupleTypes]}]
+  [:p "This attribute has " many-abbr " values, " heterogeneous-abbr " " tuples-abbr " of length "
+   (count tupleTypes) ". Each tuple is of type "
+   [heterogeneous-types-list tupleTypes]])
+(defmethod p-heterogeneous :default [{:keys [db/tupleTypes]}]
+  [:p "This attribute has " one-abbr " value, a " heterogeneous-abbr " " tuple-abbr " of length "
+   (count tupleTypes) ". The tuple is of type "
+   [heterogeneous-types-list tupleTypes]])
 
-;; scalars
-(defmethod p [:db.cardinality/one ::default ::default] [{:keys [db/valueType]}]
-  [:p "This attribute has " one-abbr " " [keyword-abbr valueType] " value."])
-(defmethod p [:db.cardinality/many ::default ::default] [{:keys [db/valueType]}]
-  [:p "This attribute has " many-abbr " " [keyword-abbr valueType] " values."])
+(defmethod p-basic :db.cardinality/many [{:keys [db/valueType]}]
+  [:p "This attribute " (has-or-references valueType) " " many-abbr " " [links-or-abbr valueType] " values."])
+(defmethod p-basic :default [{:keys [db/valueType]}]
+  [:p "This attribute " (has-or-references valueType) " " one-abbr " " [links-or-abbr valueType] " value."])
 
-(defmulti shorthand-span dispatch)
+(defmulti shorthand-span :db/valueType)
 
-;; composite tuples
-(defmethod shorthand-span [:db.type.tuple/composite ::default] [attribute]
-  [tuple-attrs-list attribute])
+(defmethod shorthand-span :db.type/tuple.composite [{:keys [db/tupleAttrs]}]
+  [tuple-attrs-list tupleAttrs])
 
-;; homogeneous tuples
-(defmethod shorthand-span [:db.type.tuple/homogeneous :referred] [{:keys [db.schema/references]}]
-  [angle-span [:span [util/coll-links references] " {2,}"]])
-(defmethod shorthand-span [:db.type.tuple/homogeneous ::default] [{:keys [db/tupleType]}]
-  [angle-span [:span [keyword-abbr tupleType] " {2,}"]])
+(defmethod shorthand-span :db.type/tuple.homogeneous [{:keys [db/tupleType]}]
+  [angle-span [:span [links-or-abbr tupleType] " {2,}"]])
 
-;; heterogeneous tuples
-(defmethod shorthand-span [:db.type.tuple/heterogeneous ::default] [attribute]
-  [heterogeneous-types-list attribute])
+(defmethod shorthand-span :db.type/tuple.heterogeneous [{:keys [db/tupleTypes]}]
+  [heterogeneous-types-list tupleTypes])
 
-;; references
-(defmethod shorthand-span [:db.type/ref :referred] [{:keys [db.schema/references]}]
-  [util/coll-links references])
-(defmethod shorthand-span [:db.type/ref ::default] [_attribute]
-  [keyword-abbr :db.type/ref])
-
-;; scalars
-(defmethod shorthand-span [::default ::default] [{:keys [db/valueType]}]
-  [keyword-abbr valueType])
+(defmethod shorthand-span :default [{:keys [db/valueType]}]
+  [pipe-links-or-abbr valueType])
 
 (defn shorthand [{:keys [db/cardinality] :as attribute}]
   (let [many? (= :db.cardinality/many cardinality)]